refactor(infra): consolidate Python tile-primitive IR nodes into tile_primitive.py
Mirror the C++ `tile_primitive.h` on the Python side: move `LambdaExpr`,
`DispatchContext`, and `TilePrimitiveCall` (plus the `normalize_const_arg`
helper) from `lambda_expr.py`, `operator/tile_primitive/dispatch_context.py`,
and `stmt.py` into a single top-level `tirx/tile_primitive.py`.

Update all direct importers and the `tirx` package re-exports. The three
classes remain available at `tvm.tirx.*`; `tvm.tirx.stmt.TilePrimitiveCall` is
gone (use `tvm.tirx.tile_primitive` or `tvm.tirx.TilePrimitiveCall`).
