+++ python/tvm/backend/cuda/operator/tile_primitive/copy/gmem_smem.py
@@ -36,7 +36,7 @@ from tvm.tirx.operator.tile_primitive.dispatcher import (
     register_dispatch,
 )
 from tvm.tirx.operator.tile_primitive.registry import DispatchContext
-from tvm.tirx.stmt import TilePrimitiveCall
+from tvm.tirx.tile_primitive import TilePrimitiveCall
 
 from ._common import (
     _TID_AXIS_FOR_SCOPE,
