+++ python/tvm/backend/cuda/operator/tile_primitive/gemm/mma_m16n8k_.py
@@ -29,7 +29,7 @@ from tvm.tirx.operator.tile_primitive import (
     predicate,
     register_dispatch,
 )
-from tvm.tirx.stmt import TilePrimitiveCall
+from tvm.tirx.tile_primitive import TilePrimitiveCall
 
 
 @dataclass(frozen=True)
