+++ python/tvm/backend/cuda/operator/tile_primitive/gemm_utils.py
@@ -20,7 +20,7 @@
 from tvm.arith.analyzer import Analyzer
 from tvm.tirx import Buffer
 from tvm.tirx.operator.tile_primitive import DispatchContext
-from tvm.tirx.stmt import TilePrimitiveCall
+from tvm.tirx.tile_primitive import TilePrimitiveCall
 
 
 def validate_gemm_op(op_call: TilePrimitiveCall, sctx: DispatchContext) -> bool:
