+++ python/tvm/backend/cuda/operator/tile_primitive/reduction/local.py
@@ -68,7 +68,7 @@ from tvm.tirx.layout import TileLayout, laneid
 from tvm.tirx.operator.tile_primitive import DispatchContext, fail
 from tvm.tirx.operator.tile_primitive.common import ReduceOpType
 from tvm.tirx.operator.tile_primitive.dispatcher import predicate, register_dispatch
-from tvm.tirx.stmt import TilePrimitiveCall
+from tvm.tirx.tile_primitive import TilePrimitiveCall
 
 from ..common import get_indices, get_st_extent
 from ..layout_utils import get_local_region, get_sublayout_from_region
