+++ python/tvm/backend/cuda/operator/tile_primitive/reduction/shared.py
@@ -63,7 +63,7 @@ from tvm.tirx import BufferRegion, PrimFunc
 from tvm.tirx.operator.tile_primitive import DispatchContext, fail
 from tvm.tirx.operator.tile_primitive.common import ReduceOpType
 from tvm.tirx.operator.tile_primitive.dispatcher import predicate, register_dispatch
-from tvm.tirx.stmt import TilePrimitiveCall
+from tvm.tirx.tile_primitive import TilePrimitiveCall
 
 from ..common import get_indices, get_st_extent, next_power_of_2
 from .utils import (
