+++ python/tvm/backend/trn/operator/tile_primitive/binary/default.py
@@ -21,7 +21,7 @@ from tvm.script import tirx as T
 from tvm.tirx import FloatImm, PrimFunc
 from tvm.tirx.operator.tile_primitive import DispatchContext, fail
 from tvm.tirx.operator.tile_primitive.common import MapOpType
-from tvm.tirx.stmt import TilePrimitiveCall
+from tvm.tirx.tile_primitive import TilePrimitiveCall
 
 from ..common import init_analyzer, nki_dim
 from ..instruction_generator import InstructionGenerator
