+++ python/tvm/backend/trn/operator/tile_primitive/gemm/default.py
@@ -31,7 +31,7 @@ from tvm.tirx.operator.tile_primitive import (
     predicate,
     register_dispatch,
 )
-from tvm.tirx.stmt import TilePrimitiveCall
+from tvm.tirx.tile_primitive import TilePrimitiveCall
 
 from ..common import init_analyzer
 from ..dim_utils import normalize_and_group
