+++ python/tvm/backend/trn/transform/private_buffer_alloc.py
@@ -19,16 +19,15 @@
 from tvm.ir import Range
 from tvm.target import Target
 from tvm.tirx.buffer import Buffer
-from tvm.tirx.operator.tile_primitive.dispatch_context import DispatchContext
 from tvm.tirx.stmt import (
     AllocBuffer,
     AttrStmt,
     For,
     SeqStmt,
     Stmt,
-    TilePrimitiveCall,
 )
 from tvm.tirx.stmt_functor import StmtMutator, StmtVisitor
+from tvm.tirx.tile_primitive import DispatchContext, TilePrimitiveCall
 from tvm.tirx.transform.common import seek_kernel_replace_point
 from tvm.tirx.transform.function_pass import prim_func_pass
