+++ python/tvm/tirx/__init__.py
@@ -44,7 +44,8 @@ from .stmt import BufferStore, AllocBuffer, AttrStmt, DeclBuffer
 from .stmt import SeqStmt
 from .stmt import IfThenElse, Evaluate, stmt_seq, stmt_list
 from .stmt import BufferRegion, MatchBufferRegion, SBlock, SBlockRealize
-from .stmt import TilePrimitiveCall, ScopeIdDefStmt
+from .stmt import ScopeIdDefStmt
+from .tile_primitive import DispatchContext, LambdaExpr, TilePrimitiveCall
 
 from .function import PrimFunc, TensorIntrin, IndexMap
 
@@ -88,7 +89,6 @@ from .generic import add, subtract, multiply
 from .exec_scope import ExecScope, ScopeIdDef
 from .layout import TileLayout, Layout, SwizzleLayout, ComposeLayout
 from .expr_functor import ExprFunctor
-from .lambda_expr import LambdaExpr
 
 from . import transform
 from . import analysis
