+++ python/tvm/tirx/operator/tile_primitive/dispatcher.py
@@ -30,9 +30,7 @@ from typing import Any
 from tvm.ir import Op
 from tvm.tirx import PrimFunc
 from tvm.tirx.operator import get_tirx_op
-from tvm.tirx.stmt import TilePrimitiveCall
-
-from .dispatch_context import DispatchContext
+from tvm.tirx.tile_primitive import DispatchContext, TilePrimitiveCall
 
 
 class DispatchFail(RuntimeError):
