+++ tests/python/tirx/operator/tile_primitive/cuda/copy_async/test_dsmem.py
@@ -35,10 +35,10 @@ from tvm.tirx import IntImm, Var
 from tvm.tirx.cuda.operator.tile_primitive.copy_async.dsmem import copy_dsmem_impl
 from tvm.tirx.exec_scope import ExecScope
 from tvm.tirx.layout import S, TileLayout
-from tvm.tirx.operator.tile_primitive.dispatch_context import DispatchContext
 from tvm.tirx.operator.tile_primitive.dispatcher import DispatchFail
 from tvm.tirx.operator.tile_primitive.ops import CopyAsync
 from tvm.tirx.stmt_functor import StmtExprVisitor
+from tvm.tirx.tile_primitive import DispatchContext
 
 
 def _make_dsmem_dispatch_call(shape, dtype, src_layout, dst_layout):
